Buyer-Template/requirements.txt
Pygments>=2.9.0
paho-mqtt>=1.5.1
idna>=3.2
requests>=2.25.1
termcolor>=1.1.0
urllib3>=1.26.6